Presentation Ready:
Working state.
Issues to be resolved later:
DB cleanup, the server is growing but there is no deletion of old data
Using React virtualization to make the body a static size to fix the stars and background.
Using multiple collections in mongo to track changes over time for things like rank and follower counts
there is a bunch of aggregation logic on the client that should be on the server
More charts.